backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
textdistance==4.5.0
python-docx==1.1.0

